removing credential cmd  - making right cli stuff
