+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "auto-timesheet",
-    "version": "0.0.33",
+    "version": "0.0.37",
     "description": "Add redmine timesheet automatically and periodically.",
     "main": "./src/bin/auto-timesheet.js",
     "bin": {
@@ -37,12 +37,13 @@
         "cookieparser": "^0.1.0",
         "editor": "^1.0.0",
         "fast-levenshtein": "^2.0.6",
-        "installed-version": "^1.0.0",
+        "jsonfile": "^3.0.1",
         "lme": "^1.5.3",
         "lru-cache": "^4.1.1",
         "request": "^2.81.0",
         "serialize-error": "^2.1.0",
-        "update-available": "^1.0.1"
+        "update-available": "^1.0.1",
+        "user-home": "^2.0.0"
     },
     "contributors": [
         "Lakshmipriya (https://twitter.com/mukundanlakshmi)"
