+++ src/bin/auto-timesheet-conf.js
@@ -1,14 +1,36 @@
 #!/usr/bin/env node
 
-const editor = require('editor');
 const path = require('path');
+const editor = require('editor');
+const confGen = require('../config/conf-template-generator');
+const home = require('user-home');
 
-editor(path.join(__dirname, '/../../credentials/config.json'), function(code) {
-	if (code !== 0) throw new Error('something went wrong while saving configuration');
+const configFile = path.join(home, '/.auto-timesheet.config.json');
 
-	console.log('done!');
-	console.log('------------------------------------------------');
-	console.log('run `auto-timesheet set-cred` to set credentials');
-	console.log('run `auto-timesheet start` to start the app');
-	console.log('------------------------------------------------enjoy');
-});
+try {
+    // check if the conf exists or not
+    require(configFile);
+    console.log('config file found: opening editor..');
+    openEditor();
+} catch (err) {
+    if (err instanceof SyntaxError) {
+        console.log('syntax err found in the config file > regenerating ...');
+    }
+    // make the file with initial values
+    else {
+        console.log('Initializing credentials ...');
+    }
+    confGen(configFile, openEditor);
+}
+
+function openEditor() {
+    console.log('opening editor');
+    return editor(configFile, function(code) {
+        if (code !== 0) throw new Error('something went wrong while saving configuration');
+
+        console.log('You\'ve edited config successfully!');
+        console.log('------------------------------------------------');
+        console.log('run `auto-timesheet start` to start the app');
+        console.log('------------------------------------------------enjoy');
+    });
+}
