+++ src/bin/auto-timesheet.js
@@ -1,11 +1,10 @@
 #!/usr/bin/env node
 
 const program = require('commander');
-const getVersion = require('installed-version');
+const version = require('../../package.json').version;
 
 program
-	.version(getVersion('auto-timesheet'))
-	.command('cred', 'set credentials')
-	.command('conf', 'configure stuff')
-	.command('start', 'start application')
-	.parse(process.argv);
+    .version(version)
+    .command('conf', 'configure stuff')
+    .command('start', 'start application')
+    .parse(process.argv);
